Day 19: Interfaces
Interfaces are a lot looser in Go than in Java, in the sense that
somebody could define an interface after a struct was written, and it
might very well automatically implement it if it implements what's
needed.

I had to modify the question a little bit to make it doable in Go.